ex10 prepared
