+++ src/ex10/main.js
@@ -1,11 +1,16 @@
 import React from "react";
 import { render } from "react-dom";
+import { createStore } from "redux";
+import { devToolsEnhancer } from "redux-devtools-extension";
 import { Provider } from "react-redux";
 
 import { Config } from "../solution/Config";
-import { App } from "./App";
 
-import { store } from "./store";
+import { App } from "./App";
+import { reducer } from "./state";
+// import { reducer } from "../solution/ex10/state";
+
+const store = createStore(reducer, devToolsEnhancer());
 
 render(
   <Config useRouter>
