+++ src/ex10/state.js
@@ -1,0 +1,10 @@
+const initialState = {};
+
+export const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case "SET_PEOPLE":
+    case "SET_PERSON":
+    default:
+      return state;
+  }
+};
